[bindings] always return value from Window self-referential getters
For whatever reason, some operations seem to invalidate the CacheAccessor unnecessarily. Rather than marking a bunch of operations as being side-effect free incorrectly, or finding another work around, it seemed the simplest solution was to simply return the v8 wrapper from the accessor. To achieve this, the wrapper object the method is invoked on is returned. The [CrossOrigin] IDL attribute remains, and the previous method of looking up the correct wrapper for a window in a detached frame is still used.

BUG=1305302

Change-Id: I2fb0f0fb9d59511debe9e009a6f3194b48ca29b5