blink: Switch root svg elements to apply all overflow properties.
This change switches the outermost root svg element (other than svg
documents where the root svg element is also the document element) to
respect paint containment and overflow-clip-margin. This aligns svg with
the implementation for other replaced elements.

This also applies overflow correctly if its different in both axis. Prior
to this change overflow was either visible in both directions or clipped
in both.

R=pdr@chromium.org
Bug:1321217
Fixed:364503
Change-Id: I9c49b16dc829d214ebd2b748638c7b4133d21509