Preserve EXIF image orientation when serializing ImageBitmap
This change fixes the serialization of ImageBitmap objects to ensure
that image orientation is preserved so that the deserialized object
behaves identically to the original.  Before this change,
structuredClone() and postMessage() were accidentally discarding the
image orientation information.

A new WPT test was added that ensures the preservation of image
orientation through a wide range of code paths that replicate
ImageBitmap objects: structured clone, postMessage, transfer via
postMessage, copy constructor.

Bug: 1332947
Change-Id: Ifeca2a46f3a10483ee57787160a07f58c4d2bf37