blink: Switch overflow overrides for replaced elements to used value.
Replaced elements support only 2 configs for overflow: visible and clip.
Currently we apply this override at style resolution based on the
element tag to identify replaced elements. This misses cases where a
an element (like img) could be rendered as a non-replaced element if
fallback to alt text is used.

Avoid this by applying the override when using the value in
layout/paint when it's known whether the element will render as a replaced element.

See discussion at: https://github.com/w3c/csswg-drafts/issues/7435.

Bug:1321217
Change-Id: Iadf653ea4a531a1ab313e2e797988d082d5ef1c1